pyproject.toml
[project]
name = "chat-app-mcp"
version = "0.0.1"
description = "De agent tools"
authors = [
    {name = "Anatoliy A Aksenov",email = "anatoliy@aksenov.org"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "langgraph (>=0.6.5,<0.7.0)",
    "fastmcp (>=2.11.3,<3.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "duckdb (>=1.3.2,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "chardet (>=5.2.0,<6.0.0)",
    "pysmb (>=1.1.22,<2.0.0)",
    "s3fs (>=2025.7.0,<2026.0.0)",
    "boto3 (>=1.39.9,<1.39.12)",
    "pandas (>=2.3.2,<3.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "pyarrow (>=21.0.0,<22.0.0)",
    "fs (>=2.4.16,<3.0.0)",
    "fs-smbfs (>=1.0.7,<2.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-gitlab (>=6.3.0,<7.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "kafka-python (>=2.2.15,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
